reorganize tests for ProjectData
Inches toward proper unit test structure and should allow easier testing
of the non-unicode metadata.csv edge case behavior.
